chore: move GetPublicKeys method to auth/service.go file
