+++ auth/service_test.go
@@ -159,3 +159,28 @@ func TestUserProfileClient_GetUserCluster(t *testing.T) {
 		})
 	}
 }
+
+func TestPublicKeys(t *testing.T) {
+
+	t.Run("valid keys", func(t *testing.T) {
+		//given
+		authService, r, err := testdoubles.NewAuthClientService("../test/data/token/auth_get_keys", "http://authservice")
+		require.NoError(t, err)
+		defer r.Stop()
+		// when
+		result, err := authService.GetPublicKeys()
+		// then
+		require.NoError(t, err)
+		assert.Len(t, result, 3)
+	})
+
+	t.Run("invalid url", func(t *testing.T) {
+		//given
+		authService, _, err := testdoubles.NewAuthClientService("", "http://google.com")
+		assert.NoError(t, err)
+		// when
+		_, err = authService.GetPublicKeys()
+		// then
+		assert.Error(t, err)
+	})
+}
